refactor: Move API related functions out of index
