+++ src/lib/api/index.jsx
@@ -70,7 +70,7 @@ const barContentComponent = (store, location) =>
  * @param  {ReduxStore} store - Store on which the API will act
  * @return {object} - Methods of the public API
  */
-export default store => {
+export const createBarAPI = store => {
   // setBar{Left,Right,Center} and <Bar{Left,Right,Center} />
   const methods = {}
   locations.forEach(location => {
@@ -91,4 +91,45 @@ export default store => {
   }
 
   return methods
+}
+
+// Handle exceptions for API before init
+const showAPIError = name => {
+  // eslint-disable-next-line no-console
+  console.error(
+    `You tried to use the CozyBar API (${name}) but the CozyBar is not initialised yet via cozy.bar.init(...).`
+  )
+}
+
+const makeProxyMethodToAPI = (exposedAPI, fnName) => {
+  return (...args) => {
+    if (exposedAPI[fnName]) {
+      return exposedAPI[fnName](...args)
+    } else {
+      showAPIError(fnName)
+    }
+  }
+}
+
+/** Creates an API that swallows error until bar is correctly initialized */
+export const createBarProxiedAPI = exposedAPI => {
+  const apiReferences = {}
+
+  locations.forEach(location => {
+    const jsAPIName = getJsApiName(location)
+    const reactAPIName = getReactApiName(location)
+    apiReferences[jsAPIName] = makeProxyMethodToAPI(exposedAPI, jsAPIName)
+    apiReferences[reactAPIName] = props => {
+      if (exposedAPI[reactAPIName]) {
+        return React.createElement(exposedAPI[reactAPIName], props)
+      } else {
+        showAPIError(reactAPIName)
+      }
+    }
+  })
+
+  for (let fnName of ['setLocale', 'setTheme']) {
+    apiReferences[fnName] = makeProxyMethodToAPI(exposedAPI, fnName)
+  }
+  return apiReferences
 }
